+++ crates/render/src/plane.rs
@@ -1,0 +1,6 @@
+#[repr(C)]
+#[derive(Copy, Clone, Debug, bytemuck::Pod, bytemuck::Zeroable)]
+pub struct Plane {
+    pub normal: [f32; 3],
+    pub height: f32,
+}
