+++ crates/render/src/run.rs
@@ -1,0 +1,52 @@
+use anyhow::{Context, Result};
+use winit::event::{Event, WindowEvent};
+use winit::event_loop::EventLoop;
+use winit::window::WindowBuilder;
+use std::sync::Arc;
+use crate::state::State;
+
+pub fn run() -> Result<()> {
+    let event_loop = EventLoop::new().context("failed to create event loop")?;
+    let window = Arc::new(WindowBuilder::new()
+        .with_title("Arena Renderer")
+        .build(&event_loop)
+        .context("failed to create window")?);
+    
+    let mut state = pollster::block_on(State::new(window.clone()))?;
+
+    event_loop.run(move |event, elwt| {
+        match event {
+            Event::WindowEvent {
+                ref event,
+                window_id,
+            } if window_id == state.window().id() => {
+                if !state.input(event) {
+                    match event {
+                        WindowEvent::CloseRequested => elwt.exit(),
+                        WindowEvent::Resized(physical_size) => {
+                            state.resize(*physical_size);
+                        }
+                        WindowEvent::RedrawRequested => {
+                            state.update();
+                            match state.render() {
+                                Ok(_) => {}
+                                // Reconfigure the surface if lost
+                                Err(wgpu::SurfaceError::Lost) => state.resize(state.window.inner_size()),
+                                // The system is out of memory, we should probably quit
+                                Err(wgpu::SurfaceError::OutOfMemory) => elwt.exit(),
+                                // All other errors (Outdated, Timeout) should be resolved by the next frame
+                                Err(e) => eprintln!("{:?}", e),
+                            }
+                        }
+                        _ => {}
+                    }
+                }
+            }
+            Event::AboutToWait => {
+                state.window().request_redraw();
+            }
+            _ => {}
+        }
+    })?;
+    Ok(())
+}
